Refactor: move assertions to ccache_core
