+++ src/Hash.cpp
@@ -20,7 +20,7 @@
 
 #include "Fd.hpp"
 #include "Logging.hpp"
-#include "fmtmacros.hpp"
+#include "core/fmtmacros.hpp"
 
 using nonstd::string_view;
