+++ src/Lockfile.cpp
@@ -20,7 +20,7 @@
 
 #include "Logging.hpp"
 #include "Util.hpp"
-#include "fmtmacros.hpp"
+#include "core/fmtmacros.hpp"
 
 #ifdef _WIN32
 #  include "Win32Util.hpp"
