+++ src/Logging.hpp
@@ -20,7 +20,7 @@
 
 #include "core/system.hpp"
 
-#include "FormatNonstdStringView.hpp"
+#include "core/FormatNonstdStringView.hpp"
 
 #include "third_party/fmt/core.h"
 #include "third_party/fmt/format.h"
