+++ src/ProgressBar.cpp
@@ -18,7 +18,7 @@
 
 #include "ProgressBar.hpp"
 
-#include "fmtmacros.hpp"
+#include "core/fmtmacros.hpp"
 
 #include "third_party/fmt/core.h"
