+++ src/ccache.cpp
@@ -28,7 +28,6 @@
 #include "Fd.hpp"
 #include "File.hpp"
 #include "Finalizer.hpp"
-#include "FormatNonstdStringView.hpp"
 #include "Hash.hpp"
 #include "Lockfile.hpp"
 #include "Logging.hpp"
@@ -48,9 +47,10 @@
 #include "cleanup.hpp"
 #include "compopt.hpp"
 #include "compress.hpp"
+#include "core/FormatNonstdStringView.hpp"
+#include "core/fmtmacros.hpp"
 #include "exceptions.hpp"
 #include "execute.hpp"
-#include "fmtmacros.hpp"
 #include "hashutil.hpp"
 #include "language.hpp"
