+++ src/core/CMakeLists.txt
@@ -1,3 +1,11 @@
-add_library(ccache_core INTERFACE)
+add_library(ccache_core
+  src/assertions.cpp
+  src/Util.cpp
+)
 
-target_include_directories(ccache_core INTERFACE include)
+target_include_directories(ccache_core PUBLIC
+  include
+  ${CMAKE_BINARY_DIR} # for config.h
+)
+
+target_link_libraries(ccache_core PUBLIC third_party_lib)
