+++ src/execute.cpp
@@ -27,7 +27,7 @@
 #include "Stat.hpp"
 #include "TemporaryFile.hpp"
 #include "Util.hpp"
-#include "fmtmacros.hpp"
+#include "core/fmtmacros.hpp"
 
 #ifdef _WIN32
 #  include "Win32Util.hpp"
