+++ unittest/test_FormatNonstdStringView.cpp
@@ -16,7 +16,7 @@
 // this program; if not, write to the Free Software Foundation, Inc., 51
 // Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#include "FormatNonstdStringView.hpp"
+#include "core/FormatNonstdStringView.hpp"
 
 #include "third_party/doctest.h"
