+++ unittest/test_Statistics.cpp
@@ -18,8 +18,8 @@
 
 #include "../src/Statistics.hpp"
 #include "../src/Util.hpp"
-#include "../src/fmtmacros.hpp"
 #include "TestUtil.hpp"
+#include "core/fmtmacros.hpp"
 
 #include "third_party/doctest.h"
