+++ unittest/test_argprocessing.cpp
@@ -21,9 +21,9 @@
 #include "../src/Context.hpp"
 #include "../src/Statistics.hpp"
 #include "../src/Util.hpp"
-#include "../src/fmtmacros.hpp"
 #include "TestUtil.hpp"
 #include "argprocessing.hpp"
+#include "core/fmtmacros.hpp"
 
 #include "third_party/doctest.h"
