+++ unittest/test_ccache.cpp
@@ -18,8 +18,8 @@
 
 #include "../src/Context.hpp"
 #include "../src/ccache.hpp"
-#include "../src/fmtmacros.hpp"
 #include "TestUtil.hpp"
+#include "core/fmtmacros.hpp"
 
 #include "third_party/doctest.h"
 #include "third_party/nonstd/optional.hpp"
